requirements.txt
attrs>=19.1.0
dtw==1.4.0
fastdtw==0.3.4
gym==0.10.9
lmdb
msgpack_numpy
numpy
torch>=1.3.1
torchvision==0.2.2.post3
tqdm>=4.0.0
yacs>=0.1.5
dgl==0.4.0
Pillow==6.1
pynvml
transformers
pytorch-lightning